Revert "make rate a floating point parameter"
This reverts commit d21cc4a8256c9910d66177e9ba68fa6167920299.

+++ TokenBucket.h
@@ -36,7 +36,7 @@ public:
   }
 
   /* rate: average number of tokens per second */
-  TokenBucket(const double rate, const uint64_t burstSize) {
+  TokenBucket(const uint64_t rate, const uint64_t burstSize) {
     timePerToken_ = 1000000 / rate;
     timePerBurst_ = burstSize * timePerToken_;
   }
